+++ src/admin-program-settings/program-settings.controller.js
@@ -28,9 +28,9 @@
         .module('admin-program-settings')
         .controller('ProgramSettingsController', ProgramSettingsController);
 
-    ProgramSettingsController.$inject = ['$state', 'program', 'programService', 'confirmService', 'notificationService', 'loadingModalService'];
+    ProgramSettingsController.$inject = ['$stateParams','$state', 'program', 'programService', 'confirmService', 'notificationService', 'loadingModalService'];
 
-    function ProgramSettingsController($state, program, programService, confirmService, notificationService, loadingModalService) {
+    function ProgramSettingsController($stateParams,$state, program, programService, confirmService, notificationService, loadingModalService) {
 
         var vm = this;
 
@@ -60,7 +60,9 @@
                 loadingModalService.open();
                 programService.update(vm.program).then(function() {
                     notificationService.success('adminProgramSettings.programSettingsSavedSuccessfully');
-                    $state.go('openlmis.administration.programs');
+                    $state.go('openlmis.administration.programs',$stateParams,{
+                        reload:true
+                    });
                 }, function() {
                     notificationService.error('adminProgramSettings.programSettingsSaveFailed');
                 }).finally(loadingModalService.close);